requirements.txt
tensorflow
opencv-python
pyautogui
stockfish
python-chess
Pillow
